Implemented getReviewsByUser() and getReviewsByCourse()

+++ src/main/java/edu/virginia/sde/reviews/Review.java
@@ -6,11 +6,11 @@ import java.util.Date;
 public class Review {
     private final int userID;
     private int courseID;
-    private double rating;
+    private int rating;  // should be 1-5
     private String comment;
     private Timestamp timestamp;
 
-    public Review(int userID, int courseID, double rating, String comment) {
+    public Review(int userID, int courseID, int rating, String comment) {
         this.userID = userID;
         this.courseID = courseID;
         this.rating = rating;
@@ -20,7 +20,7 @@ public class Review {
     }
 
     //This constructor is used in DatabaseDriver when returning a list of all reviews in the database
-    public Review(int userID, int courseID, double rating, String comment, Timestamp timestamp) {
+    public Review(int userID, int courseID, int rating, String comment, Timestamp timestamp) {
         this.userID = userID;
         this.courseID = courseID;
         this.rating = rating;
@@ -38,7 +38,7 @@ public class Review {
         return userID;
     }
 
-    public double getRating() {
+    public int getRating() {
         return rating;
     }
